# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: 90StorZon API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: tools
    description: Tool usage tracking
  - name: blog
    description: Blog operations
  - name: shop
    description: Shop operations
  - name: analytics
    description: Analytics operations
  - name: admin
    description: Admin operations
  - name: contact
    description: Contact operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current logged in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tools/track:
    post:
      operationId: trackToolUsage
      tags: [tools]
      summary: Track tool usage event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackToolBody"
      responses:
        "200":
          description: Tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /tools/stats:
    get:
      operationId: getToolStats
      tags: [tools]
      summary: Get tool usage stats
      responses:
        "200":
          description: Tool stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolStat"

  /blog/posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List all blog posts
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"

  /blog/posts/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /blog/posts/by-id/{id}:
    put:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post by ID (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBlogPostBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post by ID (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /shop/products:
    get:
      operationId: listProducts
      tags: [shop]
      summary: List all products
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [shop]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /shop/products/{id}:
    get:
      operationId: getProduct
      tags: [shop]
      summary: Get product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [shop]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [shop]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /shop/products/{id}/click:
    post:
      operationId: trackAffiliateClick
      tags: [shop]
      summary: Track an affiliate link click
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /shop/affiliate-stats:
    get:
      operationId: getAffiliateStats
      tags: [shop]
      summary: Get affiliate click stats (admin)
      responses:
        "200":
          description: Affiliate stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AffiliateStat"

  /analytics/track:
    post:
      operationId: trackPageView
      tags: [analytics]
      summary: Track a page view
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackPageViewBody"
      responses:
        "200":
          description: Tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get analytics summary (admin)
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"

  /analytics/daily:
    get:
      operationId: getDailyAnalytics
      tags: [analytics]
      summary: Get daily page view counts (admin)
      responses:
        "200":
          description: Daily analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyStat"

  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /admin/users/{id}:
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactBody"
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required: [name, email, password]

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user, message]

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, isAdmin, createdAt]

    TrackToolBody:
      type: object
      properties:
        tool:
          type: string
      required: [tool]

    ToolStat:
      type: object
      properties:
        tool:
          type: string
        count:
          type: integer
      required: [tool, count]

    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, slug, excerpt, content, category, createdAt, updatedAt]

    CreateBlogPostBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
      required: [title, slug, excerpt, content, category]

    UpdateBlogPostBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        externalLink:
          type: string
        featured:
          type: boolean
        clickCount:
          type: integer
      required: [id, name, description, price, category, featured]

    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        externalLink:
          type: string
        featured:
          type: boolean
      required: [name, description, price, category]

    AffiliateStat:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        count:
          type: integer
      required: [productId, productName, count]

    TrackPageViewBody:
      type: object
      properties:
        page:
          type: string
        referrer:
          type: string
      required: [page]

    AnalyticsSummary:
      type: object
      properties:
        totalPageViews:
          type: integer
        uniquePages:
          type: integer
        topPages:
          type: array
          items:
            type: object
            properties:
              page:
                type: string
              count:
                type: integer
            required: [page, count]
        referrers:
          type: array
          items:
            type: object
            properties:
              referrer:
                type: string
              count:
                type: integer
            required: [referrer, count]
      required: [totalPageViews, uniquePages, topPages, referrers]

    DailyStat:
      type: object
      properties:
        date:
          type: string
        count:
          type: integer
      required: [date, count]

    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalPageViews:
          type: integer
        totalToolUsage:
          type: integer
        recentUsers:
          type: array
          items:
            $ref: "#/components/schemas/User"
        toolStats:
          type: array
          items:
            $ref: "#/components/schemas/ToolStat"
      required: [totalUsers, totalPageViews, totalToolUsage, recentUsers, toolStats]

    ContactBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        message:
          type: string
      required: [name, email, message]