# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 90StorZon API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ToolStat } from "./toolStat";
import type { User } from "./user";

export interface AdminStats {
  totalUsers: number;
  totalPageViews: number;
  totalToolUsage: number;
  recentUsers: User[];
  toolStats: ToolStat[];
}