# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 90StorZon API specification
 * OpenAPI spec version: 0.2.0
 */
import type { AnalyticsSummaryReferrersItem } from "./analyticsSummaryReferrersItem";
import type { AnalyticsSummaryTopPagesItem } from "./analyticsSummaryTopPagesItem";

export interface AnalyticsSummary {
  totalPageViews: number;
  uniquePages: number;
  topPages: AnalyticsSummaryTopPagesItem[];
  referrers: AnalyticsSummaryReferrersItem[];
}