# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 90StorZon API specification
 * OpenAPI spec version: 0.2.0
 */
import type { User } from "./user";

export interface AuthResponse {
  user: User;
  message: string;
}