# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 90StorZon API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreateProductBody {
  name: string;
  description: string;
  price: number;
  imageUrl?: string;
  category: string;
  externalLink?: string;
  featured?: boolean;
}