# lib/api-zod/src/generated/types/loginBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 90StorZon API specification
 * OpenAPI spec version: 0.2.0
 */

export interface LoginBody {
  email: string;
  password: string;
}